elt-monorepo-master/pscsSourceLoader/requirements.txt
-e ../shared_libs
PyYAML
python-dotenv
pandas
polars
